# packages/protocol/src/generated/companion.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "latentPresence companion",
    "version": "1",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "paths": {
...
    "/db/embedding-models": {
      "put": {
        "operationId": "dbRegisterEmbeddingModel",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbRegisterEmbeddingModelResponse"
                }
              }
            }
          },
          "default": {
            "description": "Every failure, one shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbRegisterEmbeddingModelRequest"
              }
            }
          }
        }
      }
    },
...
    "/db/blocks": {
      "get": {
        "operationId": "dbReadBlocks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbReadBlocksResponse"
                }
              }
            }
          },
          "default": {
            "description": "Every failure, one shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "characterId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128
            }
          }
        ]
      },
      "put": {
        "operationId": "dbWriteBlock",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DbWriteBlockResponse"
                }
              }
            }
          },
          "default": {
            "description": "Every failure, one shape",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanionError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DbWriteBlockRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompanionError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "not_found",
                  "conflict",
                  "unavailable",
                  "database",
                  "upstream",
                  "internal"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "anyOf": [
                  {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {
                      "$ref": "#/$defs/__schema0"
                    }
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "$defs": {
          "__schema0": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/$defs/__schema0"
                }
              },
              {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "$ref": "#/$defs/__schema0"
                }
              }
            ]
          }
        }
      },
...
      "DbRegisterEmbeddingModelResponse": {
        "type": "object",
        "properties": {
          "modelId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "collections": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "collection": {
                  "type": "string",
                  "enum": [
                    "turns",
                    "facts",
                    "chunks"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "building",
                    "active",
                    "retired"
                  ]
                }
              },
              "required": [
                "collection",
                "status"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "modelId",
          "collections"
        ],
        "additionalProperties": false
      },
      "DbRegisterEmbeddingModelRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "model": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "dimensions": {
            "type": "integer",
            "minimum": 1,
            "maximum": 16383
          }
        },
        "required": [
          "provider",
          "model",
          "dimensions"
        ]
      },
...
      "DbReadBlocksResponse": {
        "type": "object",
        "properties": {
          "blocks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "characterId": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                "name": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 64
                },
                "content": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                },
                "editableByCharacter": {
                  "type": "boolean"
                }
              },
              "required": [
                "characterId",
                "name",
                "content",
                "updatedAt",
                "editableByCharacter"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "blocks"
        ],
        "additionalProperties": false
      },
      "DbWriteBlockResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "ok"
        ],
        "additionalProperties": false
      },
      "DbWriteBlockRequest": {
        "type": "object",
        "properties": {
          "characterId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "content": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "editableByCharacter": {
            "type": "boolean"
          }
        },
        "required": [
          "characterId",
          "name",
          "content",
          "updatedAt",
          "editableByCharacter"
        ]
      },
...
    }
  }
}